GIX-2152: ckETH not selected in governance (#3919)
# Motivation

ckETH doesn't have neurons nor proposals. It should not be visible and
the page should default to NNS.

In this PR, the page defaults to NNS, but the token is still selectable
in the Token Selector. That will come in another PR.

# Changes

* Edit `selectedUniverseIdStore` to take into account the
`icrcCanistersStore` when using NNS fallback or not.

# Tests

* Add tests for new functionality.

# Todos

- [ ] Add entry to changelog (if necessary).

Not yet necessary.