GIX-2154: Set mockCkETHToken to 18 decimals (#3935)
# Motivation

`TokenAmount` only accepts 8 decimals so to be able to upgrade ic-js, we
set the number of decimals on `mockCkETHToken` temporarily to 8.
Now we want to set it back to 18 which is allowed by using
`TokenAmountV2`.
This means we have to accept `TokenAmountV2` in many places.

This PR is not enough to make ENABLE_CKETH work again, but it's big
enough already.

# Changes

1. Make some components accept `TokenAmountV2` and some component both
`TokenAmount | TokenAmountV2` to limit the spread of the change.
2. Add `formatTokenV2` which accepts both `TokenAmount | TokenAmountV2`
but truncates decimals > 8 (for now?) and then calls `formatToken` with
a correctly scaled amount.
3. Add `toTokenAmountV2` which accepts both `TokenAmount |
TokenAmountV2` and return `TokenAmountV2`.
4. Make `numberToUlps` use `TokenAmountV2` as commented in a TODO.

# Tests

1. Updated unit tests that failed because of the changes.
2. Added tests for `formatTokenV2` and `toTokenAmountV2`.
3. Enabled tests for `numberToUlps` that were previously failing.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary