fix: Specify which flavour to update (#3947)
# Motivation
When the nns-dapp canister API changes, there are checks to make sure
that the change is intentional. There is a "golden file" that stores the
expected API for each flavour of build. If there is a mismatch in a test
build, the check prints:

```
Exports do not match
If this looks correct, run 'scripts/nns-dapp/test-exports --wasm nns-dapp_test.wasm.gz --update-golden' to update the golden file
```

That looks helpful. But what is not helpful is that the above command
updates the production golden file with the test API.

# Changes
- In the suggested command, specify which flavour of golden file to
update.

# Tests
The message is now:
```
Exports do not match
If this looks correct, run 'scripts/nns-dapp/test-exports --wasm nns-dapp_test.wasm.gz --update-golden --flavour test' to update the golden file
```
and it does update the intended flavour of golden file.

# Todos

- [x] Add entry to changelog (if necessary).