ckBTC: Merge Approve and Burn transactions into one (#3974)
# Motivation

When withdrawing ckBTC to BTC with ICRC-2 we first make an Approve
transaction and then the minter makes a Burn transaction. Since this is
part of a single process, we want to render it as a single transaction
in the wallet.

# Changes

1. Pass both the previous `IcrcTransactionWithId` and previously returns
`UiTransaction` to `mapCkbtcTransaction`. Note that transactions are
mapped in reverse chronological order, so the previous transaction
happened later in time.
2. In `mapCkbtcTransaction`, if the current transaction is an Approve
transaction, check if it was followed by a Burn transaction, and if so,
add the fee to the amount burned and remove the Approve transaction from
the list.

# Tests

1. Added a test for `mapCkbtcTransaction` with a transaction that should
be merged and with one that shouldn't be merged.
2. Added a test for `CkBTCTransactionsList`.
3. Added a convenience method to create Approve transactions.

Before:
<img width="780" alt="image"
src="https://github.com/dfinity/nns-dapp/assets/122978264/6f8018b0-fac8-4938-b01c-cab5de904b80">

After:
<img width="776" alt="image"
src="https://github.com/dfinity/nns-dapp/assets/122978264/c8fa02af-d302-475d-95fc-118e34daaae5">

# Todos

- [x] Add entry to changelog (if necessary).