GIX-2171: Use Tokens in "/" (#3985)
# Motivation

When Tokens is enabled, the home page should be the tokens page.

The current home page is the one in
`routes/(app)/(u)/(accounts)/+page.svelte`.

In this PR, I change the logic of this component to render the Tokens
route instead of the current Accounts route.

In another PR I'll move the layout and this home page to
`routes/(app)/(home)`.

# Changes

* Change logic in `(app)/(u)/(accounts)/+page.svelte` to render either
the Accounts route or the Tokens route.

# Tests

* New test for this page to check which component is rendered depending
on the flag.

# Todos

- [ ] Add entry to changelog (if necessary).

Not yet necessary.