Fix: TokensTable actions triggering link (#4010)
# Motivation

There was a navigation and full refresh when the user clicked directly
in one of the action buttons.

We need to prevent the default behavior of the link as well as stop
propagation.

# Changes

* Add `preventDefault` directive to ReceiveButton and SendButton.

The GoToDetailButton will be refactoring into a single span with no
clicking functionality because we should rely on the rowHref instead.

# Tests

* Test that when clicking Send or Receive, the default is prevented.

# Todos

- [ ] Add entry to changelog (if necessary).

Not necessary.