GIX-2062: Refactor tokensListBaseStore (#4012)
# Motivation

Add loading state to the balance of the Tokens Table.

In this PR, I refactor the `tokensListBaseStore` to return the type
`UserTokenBase` instead of `UserTokenData`.

Yet, the outcome of the related `tokensListUserStore` and
`tokensListVisitorsStore` doesn't change. As can be assessed by the lack
of changes in their spec files.

# Changes

* `tokensListBaseStore` return a simple `UserTokenBase`.
* `tokensListUserStore` converts from `UserTokenBase` to
`UserTokenData`. I have kept the same logic as before. In a following PR
I will add the loading logic. I have also merged the helpers together to
avoid handling intermediate types.
* `tokensListVisitorStore` converts from `UserTokenBase` to
`UserTokenData`. I have kept the same logic as before. In a following PR
I will add the loading logic.

# Tests

* Adapt test for `tokensListBaseStore`.
* Change the user token mocks to be of type `UserTokenBase`

# Todos

- [ ] Add entry to changelog (if necessary).

Not necessary.