GIX-2180: Add accountIdentifier UserTokenData (#4060)
# Motivation

To open the TransactionModal or ReceiveModal in the ICP tokens table, I
need to know which account to use based on the row clicked.

I could use the title, but that is fragile because the title might
change and not be the name of the account.

Instead, I added an optional field to UserTokenData: accounIdentifier.

# Changes

* New field `accountIdentifier` in `UserTokenData`.
* Populate the new field in `icpTokensListUser`.
* Populate the new field in `tokensListUserStore`. I changed the
implementation to read all the universes accounts instead of the
balance. I needed it to get the account identifier.
* Use function overloading in helper `sumAccounts`. Convenient to make
sure the return is a `bigint` if there are accounts. I also changed the
functionality slightly to return 0 when the list is empty. That made
more sense to me.

# Tests

* Adapt the tests expectation for `icpTokensListUser`.
* Adapt the tests expectation for `tokensListUserStore`.
* Add tests `sumAccounts` edge cases.

# Todos

- [ ] Add entry to changelog (if necessary).

Not yet necessary.