Await Playwright page.evaluate (#4066)
# Motivation

I noticed a CI run where the token symbols were not masked out in the
screenshot.
And I noticed we don't await the `page.evaluate` call that replaces the
content.
I don't know if this is the reason but awaiting the call is the right
thing to do.

# Changes

Await the Playwright `page.evaluate` call in `replaceContent`.

# Tests

test only

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary