Service method to load BTC retrieval statuses (#4076)
# Motivation

We want to render failed and pending "Sending BTC" transactions as such.
These statuses are not part of the transactions themselves but come from
the ckBTC minter.

# Changes

Add a service method `loadRetrieveBtcStatuses` which loads BTC retrieval
statuses and puts them in the store.

# Tests

Unit tests were added.
Tested manually in another branch.

# Todos

- [ ] Add entry to changelog (if necessary).
will add when it's used