Take retrieval status into account in mapCkbtcTransaction (#4084)
# Motivation

We want to render failed and pending "Sending BTC" transactions as such.
These statuses are not part of the transactions themselves but come from
the ckBTC minter and then put in a store.

In this PR we change the mapping function to take the status into
account when creation the `UiTransaction`.

# Changes

1. Add a new `isFailed` field to `UiTransaction` to indicate a failed
transaction.
2. Add an optional parameter to `mapCkbtcTransaction` to pass the
retrieval status if there is one.
3. If there is a status, output a pending or failed transaction if
appropriate.
4. Add i18n strings for the transaction headlines.

# Tests

Unit tests added.
Tested manually in another branch with more changes.

# Todos

- [ ] Add entry to changelog (if necessary).
not yet