Pass BTC retrieval statuses when mapping ckBTC transactions (#4108)
# Motivation

We want to render failed and pending "Sending BTC" transactions as such.
These statuses are not part of the transactions themselves but come from
the ckBTC minter and will be put in a store (in a future PR).

In this PR we take the statuses from the store and pass them to
`mapCkbtcTransactions` so they can be used to map the transactions.

# Changes

In `CkBTCTransactionsList.svelte`, pass statuses from
`ckbtcRetrieveBtcStatusesStore` to `mapCkbtcTransactions`.

# Tests

Added a unit test where we put a status in the store and verify that it
renders the corresponding transaction according to the status.

Tested manually in another branch.

# Todos

- [ ] Add entry to changelog (if necessary).
will add when we fill the store