GIX-2187: Redirect from SNS wallet to accounts for wrong account identifier (#4113)
# Motivation

When you go to a wallet page without an account identifier in the URL,
this is invalid and you are redirected to the accounts page.
But not for SNS wallets. There you just get an infinite loading spinner.
For consistency, SNS wallets should also redirect to the accounts page.

# Changes

In `SnsWallet.svelte`, when the account identifier does not match any of
the accounts, go back to the accounts page of the same universe.
This is similar to what
[IcrcWalletPage.svelte](https://github.com/dfinity/nns-dapp/blob/fd4ce83c1fc11801b43041359bdebfb58b5d3b6d/frontend/src/lib/components/accounts/IcrcWalletPage.svelte#L74)
does (although there it doesn't keep the universe which is a bug that
should be fixed in another PR).

# Tests

Added unit tests for missing, invalid and valid account identifier.
Tested manually.

# Todos

- [x] Add entry to changelog (if necessary).